refactor(vat-data): Merge kind-utils.js into exo-utils.js
