+++ packages/vat-data/src/index.js
@@ -35,7 +35,7 @@ export {
   prepareExoClass,
   prepareExoClassKit,
   prepareExo,
-  // deorecated
+  // deprecated
   prepareSingleton,
 } from './exo-utils.js';
 
@@ -45,10 +45,12 @@ export {
 
 // //////////////////////////// deprecated /////////////////////////////////////
 
+/**
+ * @deprecated Use Exos/ExoClasses instead of Kinds
+ */
 export {
-  // deprecated
   ignoreContext,
   provideKindHandle,
   prepareKind,
   prepareKindMulti,
-} from './kind-utils.js';
+} from './exo-utils.js';
